refactor: relocate UUID and NodeId out of litegraph
Move the canonical UUID utilities and the `NodeId` alias up out of
`src/lib/litegraph/` so non-litegraph code can reference them without
crossing the litegraph layer boundary.

- `src/lib/litegraph/src/utils/uuid.ts` → `src/utils/uuid.ts` (whole
  file: `UUID`, `zeroUuid`, `createUuidv4`).
- `NodeId` definition moves from `src/lib/litegraph/src/LGraphNode.ts`
  to `src/world/entityIds.ts`. `LGraphNode.ts` re-exports it for
  backward compatibility, and the litegraph barrel still re-exports
  `createUuidv4` / `UUID` so the package's public surface is unchanged.
- All 22 importers updated to use `@/utils/uuid`.
- Drop the two `import-x/no-restricted-paths` eslint disables in
  `src/world/entityIds.ts` that were waiting on these moves.

No behavior change.

Amp-Thread-ID: https://ampcode.com/threads/T-019e8580-a262-71ac-9361-7d80aa73f5bb
