+++ src/lib/litegraph/src/subgraph/EmptySubgraphInput.ts
@@ -3,7 +3,7 @@ import type { LLink } from '@/lib/litegraph/src/LLink'
 import type { RerouteId } from '@/lib/litegraph/src/Reroute'
 import type { INodeInputSlot, Point } from '@/lib/litegraph/src/interfaces'
 import { nextUniqueName } from '@/lib/litegraph/src/strings'
-import { zeroUuid } from '@/lib/litegraph/src/utils/uuid'
+import { zeroUuid } from '@/utils/uuid'
 
 import { SubgraphInput } from './SubgraphInput'
 import type { SubgraphInputNode } from './SubgraphInputNode'
