+++ src/lib/litegraph/src/types/serialisation.ts
@@ -1,4 +1,4 @@
-import type { UUID } from '@/lib/litegraph/src/utils/uuid'
+import type { UUID } from '@/utils/uuid'
 
 import type {
   LGraphConfig,
